Add a bunch of deps, try to fix shadow jar, also cried a lot

+++ src/main/java/com/diamonddagger590/mccore/task/core/CancellableCoreTask.java
@@ -12,7 +12,7 @@ public abstract class CancellableCoreTask extends RepeatableCoreTask {
 
     protected boolean cancelled;
 
-    public CancellableCoreTask(@NotNull Plugin plugin, long taskDelay, long taskFrequency) {
+    public CancellableCoreTask(@NotNull Plugin plugin, double taskDelay, double taskFrequency) {
         super(plugin, taskDelay, taskFrequency);
     }
